refactor: split mixed-concern runtime modules
