+++ yoyopy/cli/remote/ops.py
@@ -12,293 +12,44 @@ import subprocess
 import sys
 import tarfile
 import tempfile
-from dataclasses import dataclass
 from pathlib import Path
 from typing import Annotated, Optional, Sequence
 
 import typer
-import yaml
 
 from yoyopy.cli.common import REPO_ROOT
+from yoyopy.cli.remote.config import (
+    DEFAULT_PI_PROJECT_DIR,
+    DEPLOY_CONFIG_PATH,
+    LOCAL_DEPLOY_CONFIG_PATH,
+    PiDeployConfig,
+    RemoteConfig,
+    load_pi_deploy_config,
+    pi_deploy_config_to_dict,
+    resolve_remote_config,
+)
+from yoyopy.cli.remote.transport import (
+    quote_remote_project_dir,
+    run_local,
+    run_local_capture,
+    run_remote,
+    run_remote_capture,
+    shell_quote,
+    validate_config,
+)
 
-# ---------------------------------------------------------------------------
-# Deploy config model
-# ---------------------------------------------------------------------------
-
-DEPLOY_CONFIG_PATH = REPO_ROOT / "deploy" / "pi-deploy.yaml"
-LOCAL_DEPLOY_CONFIG_PATH = REPO_ROOT / "deploy" / "pi-deploy.local.yaml"
-DEFAULT_PI_PROJECT_DIR = "~/YoyoPod_Core"
-
-
-@dataclass
-class RemoteConfig:
-    """Connection details for the Raspberry Pi host."""
-
-    host: str
-    user: str
-    project_dir: str
-    branch: str
-
-    @property
-    def ssh_target(self) -> str:
-        """Return the SSH target in user@host form when a user is configured."""
-        if self.user:
-            return f"{self.user}@{self.host}"
-        return self.host
-
-
-@dataclass(frozen=True)
-class PiDeployConfig:
-    """Stable runtime paths used by the Pi deploy/debugging workflow."""
-
-    host: str = ""
-    user: str = ""
-    project_dir: str = DEFAULT_PI_PROJECT_DIR
-    branch: str = "main"
-    venv: str = ".venv"
-    start_cmd: str = "python yoyopod.py"
-    kill_processes: tuple[str, ...] = ("python", "linphonec")
-    log_file: str = "logs/yoyopod.log"
-    error_log_file: str = "logs/yoyopod_errors.log"
-    pid_file: str = "/tmp/yoyopod.pid"
-    startup_marker: str = "YoyoPod starting"
-    screenshot_path: str = "/tmp/yoyopod_screenshot.png"
-    rsync_exclude: tuple[str, ...] = (
-        ".git/",
-        ".cache/",
-        "__pycache__/",
-        "*.pyc",
-        ".venv/",
-        "build/",
-        "logs/",
-        "models/",
-        "node_modules/",
-        "*.egg-info/",
-    )
-
-
-# ---------------------------------------------------------------------------
-# Config loading
-# ---------------------------------------------------------------------------
-
-
-def load_yaml_mapping(path: Path) -> dict[str, object]:
-    """Load one YAML mapping from disk."""
-    with open(path, "r", encoding="utf-8") as handle:
-        data = yaml.safe_load(handle) or {}
-    if not isinstance(data, dict):
-        raise SystemExit(f"Expected a YAML mapping in {path}")
-    return data
-
-
-def merge_pi_deploy_layers(*layers: dict[str, object]) -> dict[str, object]:
-    """Merge deploy config layers from lowest to highest precedence."""
-    merged: dict[str, object] = {}
-    for layer in layers:
-        for key, value in layer.items():
-            if value is not None:
-                merged[key] = value
-    return merged
-
-
-def _as_string_tuple(value: object, *, default: tuple[str, ...]) -> tuple[str, ...]:
-    """Normalize one YAML sequence-like value into a tuple of non-empty strings."""
-
-    if isinstance(value, str):
-        candidates: Sequence[object] = (value,)
-    elif isinstance(value, Sequence) and not isinstance(value, (bytes, bytearray)):
-        candidates = value
-    else:
-        return default
-
-    normalized = tuple(str(item).strip() for item in candidates if str(item).strip())
-    return normalized or default
-
-
-def parse_pi_deploy_config(data: dict[str, object]) -> PiDeployConfig:
-    """Normalize raw YAML data into a deploy config object."""
-    return PiDeployConfig(
-        host=str(data.get("host", "")).strip(),
-        user=str(data.get("user", "")).strip(),
-        project_dir=str(
-            data.get("project_dir", data.get("remote_dir", DEFAULT_PI_PROJECT_DIR))
-        ).strip()
-        or DEFAULT_PI_PROJECT_DIR,
-        branch=str(data.get("branch", "main")).strip() or "main",
-        venv=str(data.get("venv", ".venv")).strip() or ".venv",
-        start_cmd=str(data.get("start_cmd", "python yoyopod.py")).strip() or "python yoyopod.py",
-        kill_processes=_as_string_tuple(
-            data.get("kill_processes", ("python", "linphonec")),
-            default=("python", "linphonec"),
-        ),
-        log_file=str(data["log_file"]).strip(),
-        error_log_file=str(data["error_log_file"]).strip(),
-        pid_file=str(data["pid_file"]).strip(),
-        startup_marker=str(data["startup_marker"]).strip(),
-        screenshot_path=str(data.get("screenshot_path", "/tmp/yoyopod_screenshot.png")).strip()
-        or "/tmp/yoyopod_screenshot.png",
-        rsync_exclude=_as_string_tuple(
-            data.get(
-                "rsync_exclude",
-                (
-                    ".git/",
-                    ".cache/",
-                    "__pycache__/",
-                    "*.pyc",
-                    ".venv/",
-                    "build/",
-                    "logs/",
-                    "models/",
-                    "node_modules/",
-                    "*.egg-info/",
-                ),
-            ),
-            default=(
-                ".git/",
-                ".cache/",
-                "__pycache__/",
-                "*.pyc",
-                ".venv/",
-                "build/",
-                "logs/",
-                "models/",
-                "node_modules/",
-                "*.egg-info/",
-            ),
-        ),
-    )
-
-
-def load_pi_deploy_config(
-    *,
-    config_path: Path | None = None,
-    local_override_path: Path | None = None,
-) -> PiDeployConfig:
-    """Load the shared deploy config with an optional local override layer."""
-    base_path = config_path or DEPLOY_CONFIG_PATH
-    local_path = local_override_path or LOCAL_DEPLOY_CONFIG_PATH
-
-    merged_data = load_yaml_mapping(base_path)
-    if local_path.exists():
-        merged_data = merge_pi_deploy_layers(
-            merged_data,
-            load_yaml_mapping(local_path),
-        )
-
-    return parse_pi_deploy_config(merged_data)
-
-
-def pi_deploy_config_to_dict(config: PiDeployConfig) -> dict[str, object]:
-    """Convert one deploy config object back into a YAML-friendly mapping."""
-    return {
-        "host": config.host,
-        "user": config.user,
-        "project_dir": config.project_dir,
-        "branch": config.branch,
-        "venv": config.venv,
-        "start_cmd": config.start_cmd,
-        "kill_processes": list(config.kill_processes),
-        "log_file": config.log_file,
-        "error_log_file": config.error_log_file,
-        "pid_file": config.pid_file,
-        "startup_marker": config.startup_marker,
-        "screenshot_path": config.screenshot_path,
-        "rsync_exclude": list(config.rsync_exclude),
-    }
-
-
-# ---------------------------------------------------------------------------
-# SSH and subprocess helpers
-# ---------------------------------------------------------------------------
-
-
-def shell_quote(value: str) -> str:
-    """Shell-escape one literal value for the remote command string."""
-    return shlex.quote(value)
-
-
-def quote_remote_project_dir(project_dir: str) -> str:
-    """Quote the remote project path while preserving ``~`` expansion."""
-    if project_dir == "~":
-        return '"$HOME"'
-    if project_dir.startswith("~/"):
-        suffix = project_dir[2:].replace('"', '\\"')
-        return f'"$HOME/{suffix}"'
-    return shlex.quote(project_dir)
-
-
-def build_ssh_command(
-    config: RemoteConfig,
-    remote_command: str,
-    *,
-    tty: bool = False,
-) -> list[str]:
-    """Build one SSH command targeting the Raspberry Pi."""
-    wrapped_command = f"cd {quote_remote_project_dir(config.project_dir)} && {remote_command}"
-    ssh_command = ["ssh"]
-    if tty:
-        ssh_command.append("-t")
-    ssh_command.extend([config.ssh_target, f"bash -lc {shlex.quote(wrapped_command)}"])
-    return ssh_command
-
-
-def run_remote(config: RemoteConfig, remote_command: str, tty: bool = False) -> int:
-    """Execute one command on the Raspberry Pi via SSH."""
-    ssh_command = build_ssh_command(config, remote_command, tty=tty)
-
-    print("")
-    print(f"[pi-remote] host={config.ssh_target}")
-    print(f"[pi-remote] dir={config.project_dir}")
-    print(f"[pi-remote] cmd={remote_command}")
-    print("")
-
-    completed = subprocess.run(ssh_command, check=False)
-    return completed.returncode
-
-
-def run_remote_capture(
-    config: RemoteConfig,
-    remote_command: str,
-) -> subprocess.CompletedProcess[str]:
-    """Execute one SSH command and capture its stdout/stderr."""
-    ssh_command = build_ssh_command(config, remote_command)
-    return subprocess.run(
-        ssh_command,
-        check=False,
-        capture_output=True,
-        text=True,
-    )
-
-
-def run_local(command: Sequence[str], label: str) -> int:
-    """Execute one local command and stream its output."""
-    print("")
-    print(f"[pi-remote] local={label}")
-    print(f"[pi-remote] cmd={shlex.join(command)}")
-    print("")
-
-    completed = subprocess.run(list(command), check=False)
-    return completed.returncode
-
-
-def run_local_capture(command: Sequence[str]) -> subprocess.CompletedProcess[str]:
-    """Execute one local command and capture its stdout/stderr."""
-    return subprocess.run(
-        list(command),
-        check=False,
-        capture_output=True,
-        text=True,
-    )
-
-
-def validate_config(config: RemoteConfig) -> None:
-    """Ensure required connection details are present."""
-    if not config.host:
-        raise SystemExit(
-            "Missing Raspberry Pi host. Set it with "
-            "`uv run yoyoctl remote config edit`, "
-            "pass --host, or set YOYOPOD_PI_HOST."
-        )
+__all__ = [
+    "DEFAULT_PI_PROJECT_DIR",
+    "DEPLOY_CONFIG_PATH",
+    "LOCAL_DEPLOY_CONFIG_PATH",
+    "PiDeployConfig",
+    "RemoteConfig",
+    "_resolve_remote_config",
+    "load_pi_deploy_config",
+    "pi_deploy_config_to_dict",
+    "quote_remote_project_dir",
+    "resolve_local_validation_target",
+]
 
 
 def _activate_script_path(venv: str) -> str:
@@ -315,14 +66,9 @@ def _resolve_remote_config(
     project_dir: str,
     branch: str,
 ) -> RemoteConfig:
-    """Build a RemoteConfig from CLI option values."""
-    deploy_config = load_pi_deploy_config()
-    return RemoteConfig(
-        host=host or os.getenv("YOYOPOD_PI_HOST", deploy_config.host),
-        user=user or os.getenv("YOYOPOD_PI_USER", deploy_config.user),
-        project_dir=project_dir or os.getenv("YOYOPOD_PI_PROJECT_DIR", deploy_config.project_dir),
-        branch=branch or os.getenv("YOYOPOD_PI_BRANCH", deploy_config.branch),
-    )
+    """Backward-compatible wrapper for sibling remote CLI modules."""
+
+    return resolve_remote_config(host, user, project_dir, branch)
 
 
 def _capture_local_git(command: Sequence[str], *, action: str) -> str:
@@ -999,7 +745,7 @@ def status(
     ] = "",
 ) -> None:
     """Show remote repo, music backend, and process status."""
-    config = _resolve_remote_config(host, user, project_dir, branch)
+    config = resolve_remote_config(host, user, project_dir, branch)
     validate_config(config)
     deploy_config = load_pi_deploy_config()
     rc = run_remote(config, build_status_command(deploy_config))
@@ -1032,7 +778,7 @@ def sync(
     ] = False,
 ) -> None:
     """Sync the stable Pi checkout to one committed branch or exact commit."""
-    config = _resolve_remote_config(host, user, project_dir, branch)
+    config = resolve_remote_config(host, user, project_dir, branch)
     validate_config(config)
     rc = run_remote(config, build_sync_command(config, skip_uv_sync, target_sha=sha))
     if rc != 0:
@@ -1103,7 +849,7 @@ def validate(
 ) -> None:
     """Validate a committed branch/SHA on the Pi, then leave the app running."""
     resolved_branch, resolved_sha = resolve_local_validation_target(branch=branch, sha=sha)
-    config = _resolve_remote_config(host, user, project_dir, resolved_branch)
+    config = resolve_remote_config(host, user, project_dir, resolved_branch)
     validate_config(config)
     deploy_config = load_pi_deploy_config()
 
@@ -1184,7 +930,7 @@ def smoke(
     ] = 90.0,
 ) -> None:
     """Run the Raspberry Pi smoke validator remotely."""
-    config = _resolve_remote_config(host, user, project_dir, branch)
+    config = resolve_remote_config(host, user, project_dir, branch)
     validate_config(config)
     rc = run_remote(
         config,
@@ -1268,7 +1014,7 @@ def preflight(
     ] = 90.0,
 ) -> None:
     """Run local checks, sync the Pi, and execute the Pi smoke pass."""
-    config = _resolve_remote_config(host, user, project_dir, branch)
+    config = resolve_remote_config(host, user, project_dir, branch)
     validate_config(config)
 
     if not skip_local:
@@ -1318,7 +1064,7 @@ def restart(
     verbose: Annotated[bool, typer.Option("--verbose", help="Enable debug logging.")] = False,
 ) -> None:
     """Restart the yoyopod app on the Pi."""
-    config = _resolve_remote_config(host, user, project_dir, branch)
+    config = resolve_remote_config(host, user, project_dir, branch)
     validate_config(config)
     deploy_config = load_pi_deploy_config()
     rc = run_remote(config, build_restart_command(deploy_config))
@@ -1350,7 +1096,7 @@ def logs(
     verbose: Annotated[bool, typer.Option("--verbose", help="Enable debug logging.")] = False,
 ) -> None:
     """Tail yoyopod logs on the Pi."""
-    config = _resolve_remote_config(host, user, project_dir, branch)
+    config = resolve_remote_config(host, user, project_dir, branch)
     validate_config(config)
     deploy_config = load_pi_deploy_config()
     args = argparse.Namespace(
@@ -1389,7 +1135,7 @@ def screenshot(
     verbose: Annotated[bool, typer.Option("--verbose", help="Enable debug logging.")] = False,
 ) -> None:
     """Capture a screenshot from the Pi display."""
-    config = _resolve_remote_config(host, user, project_dir, branch)
+    config = resolve_remote_config(host, user, project_dir, branch)
     validate_config(config)
     deploy_config = load_pi_deploy_config()
     args = argparse.Namespace(readback=readback, output=output)
@@ -1417,7 +1163,7 @@ def rsync(
     verbose: Annotated[bool, typer.Option("--verbose", help="Enable debug logging.")] = False,
 ) -> None:
     """Rare-case escape hatch: mirror the local dirty working tree to the Pi."""
-    config = _resolve_remote_config(host, user, project_dir, branch)
+    config = resolve_remote_config(host, user, project_dir, branch)
     validate_config(config)
     deploy_config = load_pi_deploy_config()
     print(
@@ -1459,7 +1205,7 @@ def whisplay(
     verbose: Annotated[bool, typer.Option("--verbose", help="Enable debug logging.")] = False,
 ) -> None:
     """Run the Whisplay gesture-tuning helper remotely."""
-    config = _resolve_remote_config(host, user, project_dir, branch)
+    config = resolve_remote_config(host, user, project_dir, branch)
     validate_config(config)
     args = argparse.Namespace(
         verbose=verbose,
@@ -1500,7 +1246,7 @@ def rtc(
     verbose: Annotated[bool, typer.Option("--verbose", help="Enable debug logging.")] = False,
 ) -> None:
     """Inspect or control PiSugar RTC state remotely."""
-    config = _resolve_remote_config(host, user, project_dir, branch)
+    config = resolve_remote_config(host, user, project_dir, branch)
     validate_config(config)
     args = argparse.Namespace(
         verbose=verbose,
